+++ examples/demos/emberjs_lmd/js/routes/todosIndex_route.js
@@ -1,0 +1,7 @@
+var Ember = require('ember');
+
+module.exports = Ember.Route.extend({
+    setupController: function () {
+        this.controllerFor('todos').set('filteredTodos', this.modelFor('todos'));
+    }
+});