+++ examples/demos/emberjs_lmd/js/routes/todos_route.js
@@ -1,0 +1,7 @@
+var Ember = require('ember');
+
+module.exports = Ember.Route.extend({
+    model: function () {
+        return this.store.find('todo');
+    }
+});